move producer mock out of runs implementation
