+++ internal/events/mocks/definition.go
@@ -1,0 +1,5 @@
+//go:build mocks
+
+package mocks
+
+//go:generate mockery --output . --filename ./run_producer_mock.go 	--dir .. --name Producer
