--- internal/events/runs/mocks/definition.go
@@ -1,5 +1,0 @@
-//go:build mocks
-
-package mocks
-
-//go:generate mockery --output . --filename ./run_producer_mock.go 	--dir ../.. --name Producer
